move FamilySidebar styles to new file
